# services/api-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  version: 1.0.0
  description: |
    Единая точка входа в интернет-магазин. Все клиентские запросы идут через
    API Gateway, который проверяет JWT-токен, подставляет заголовки
    `X-User-Id` / `X-User-Roles` и проксирует вызовы в микросервисы:

    | Раздел API        | Микросервис     | Транспорт |
    |-------------------|-----------------|-----------|
    | `/auth/*`         | auth-service    | HTTP      |
    | `/users/*`        | user-service    | HTTP      |
    | `/products/*`     | product-service | HTTP      |
    | `/categories/*`   | product-service | HTTP      |
    | `/inventory/*`    | inventory-service | HTTP    |
    | `/orders/*`       | order-service   | HTTP      |
    | остатки в каталоге | inventory-service | gRPC    |

    ## Как тестировать

    1. `POST /api/v1/auth/register` — регистрация (получите письмо с токеном подтверждения).
    2. `POST /api/v1/auth/login` — вход. В ответе придёт `access_token` и cookie `refresh_token`.
    3. Нажмите **Authorize** в Swagger UI и вставьте `access_token` (без слова `Bearer`).
    4. Все защищённые эндпоинты станут доступны.

    ## Авторизация

    - Защищённые эндпоинты требуют заголовок `Authorization: Bearer <access_token>`.
    - `access_token` — короткоживущий JWT (RS256, issuer `auth-service`).
    - `refresh_token` передаётся в HttpOnly-cookie `refresh_token`
      и используется только в `POST /api/v1/auth/refresh` и `POST /api/v1/auth/logout`.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users/me:
    get:
      summary: Профиль текущего пользователя
      description: Возвращает профиль по `user_id` из access-токена, включая роли.
      tags: [users]
      operationId: getMyProfile
      responses:
        '200':
          description: Профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                email: user@example.com
                first_name: Иван
                last_name: Петров
                phone: '+79991234567'
                avatar_url: https://cdn.example.com/avatars/1.png
                roles: [ROLE_USER]
                created_at: '2024-01-15T10:30:00Z'
                updated_at: '2024-01-15T10:30:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Обновить свой профиль
      description: |
        Частичное обновление: передавайте только изменяемые поля.
        Поля, не переданные в теле запроса, остаются без изменений.
        Email изменить нельзя.
      tags: [users]
      operationId: updateMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            examples:
              partial:
                summary: Изменить только имя
                value:
                  first_name: Иван
              full:
                summary: Заполнить профиль полностью
                value:
                  first_name: Иван
                  last_name: Петров
                  phone: '+79991234567'
                  avatar_url: https://cdn.example.com/avatars/1.png
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users:
    get:
      summary: Список пользователей
      description: |
        Страница списка пользователей с offset-пагинацией.
        Доступно только пользователям с ролью `ROLE_ADMIN`.
      tags: [users]
      operationId: listUsers
      parameters:
        - name: limit
          in: query
          description: |
            Количество пользователей в ответе.
            Значения вне диапазона 1–100 игнорируются, используется 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Количество пропускаемых пользователей
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Страница списка пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'

    get:
      summary: Получить пользователя по ID
      description: Доступно только пользователям с ролью `ROLE_ADMIN`.
      tags: [users]
      operationId: getUserById
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: "`id` не является корректным UUID"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid id format
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Удалить пользователя
      description: |
        Удаляет пользователя (soft delete). Доступно только `ROLE_ADMIN`.
        Нельзя удалить самого себя.
      tags: [users]
      operationId: deleteUser
      responses:
        '204':
          description: Пользователь удалён
        '400':
          description: "`id` не является корректным UUID"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Нельзя удалить самого себя или последнего администратора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access-токен, полученный в `POST /api/v1/auth/login`
        или `POST /api/v1/auth/refresh`.

        В Swagger UI нажмите **Authorize** и вставьте только значение
        токена, без префикса `Bearer `.

        Получить токен:
        ```bash
        curl -s -X POST http://localhost:4000/api/v1/auth/login \
          -H 'Content-Type: application/json' \
          -d '{"email":"user@example.com","password":"password123"}'
        ```

        Токен подписан RS256, issuer — `auth-service`. Gateway проверяет
        подпись, срок действия и `exp`/`nbf` с небольшим допуском (leeway).
        Истёкший или подделанный токен → `401`.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: |
        HttpOnly-cookie с refresh-токеном. Выставляется автоматически
        ответами `login`, `oauth/callback` и `refresh`.

        Cookie имеет флаги `Secure` и `Path=/api/v1/auth/refresh`,
        поэтому по `http://localhost:4000` браузер её не сохранит —
        для локального тестирования задайте значение cookie вручную
        через **Authorize** в Swagger UI.
  parameters:
...
    UserIdPath:
      name: id
      in: path
      required: true
      description: ID пользователя (UUID)
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
    ErrorResponse:
      type: object
      description: Единый формат ошибки для всех эндпоинтов шлюза
      required: [error]
      properties:
        error:
          type: string
          description: Сообщение об ошибке (может быть на русском языке)
          example: Проверьте корректность данных
    UserRole:
      type: string
      description: |
        Роль пользователя:
        - `ROLE_USER` — покупатель
        - `ROLE_MANAGER` — управление товарами, остатками, заказами
        - `ROLE_ADMIN` — полный доступ, включая пользователей и категории
      enum: [ROLE_USER, ROLE_MANAGER, ROLE_ADMIN]
...
    User:
      type: object
      description: Профиль пользователя
      required: [id, email, roles, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        first_name:
          type: string
          nullable: true
          example: Иван
        last_name:
          type: string
          nullable: true
          example: Петров
        phone:
          type: string
          nullable: true
          example: '+79991234567'
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.com/avatars/1.png
        roles:
          type: array
          description: Роли пользователя
          items:
            $ref: '#/components/schemas/UserRole'
          example: [ROLE_USER]
        created_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Заполняется только для мягко удалённых пользователей
          example: null
    UpdateProfileRequest:
      type: object
      description: |
        Частичное обновление профиля. Передавайте только изменяемые поля.
        Email и роли изменить нельзя.
      properties:
        first_name:
          type: string
          example: Иван
        last_name:
          type: string
          example: Петров
        phone:
          type: string
          example: '+79991234567'
        avatar_url:
          type: string
          format: uri
          example: https://cdn.example.com/avatars/1.png
    ListUsersResponse:
      type: object
      description: Страница списка пользователей с offset-пагинацией
      required: [users, total, limit, offset]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          description: Всего пользователей в системе
          example: 42
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
...
  responses:
...
    ValidationError:
      description: Ошибка валидации тела запроса (`Проверьте корректность данных`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Проверьте корректность данных
    Unauthorized:
      description: |
        Не передан `Authorization: Bearer <token>`, токен невалиден или истёк
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: Токен не передан
              value:
                error: missing access token
            invalid:
              summary: Токен невалиден или истёк
              value:
                error: invalid or expired access token
    Forbidden:
      description: Недостаточно прав — у пользователя нет требуемой роли
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: insufficient permissions
    NotFound:
      description: Запрашиваемый ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: |
        Превышен лимит частоты запросов. Повторите через
        `X-RateLimit-Retry-After` секунд.
      headers:
        X-RateLimit-Limit:
          description: Максимальное число запросов в окне
          schema:
            type: integer
          example: 60
        X-RateLimit-Remaining:
          description: Сколько запросов осталось
          schema:
            type: integer
          example: 0
        X-RateLimit-RetryAfter:
          description: Через сколько секунд можно повторить
          schema:
            type: integer
          example: 2
        X-RateLimit-Reset:
          description: Unix-время сброса окна
          schema:
            type: integer
            format: int64
          example: 1736940000
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: too many requests
    InternalError:
      description: Внутренняя ошибка шлюза или недоступность микросервиса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error